requirements.txt
asgiref==3.7.2
click==8.1.7
Django==4.2.9
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==7.0.1
joblib==1.3.2
Markdown==3.5.2
nltk==3.8.1
pytz==2023.4
regex==2023.12.25
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
zipp==3.17.0
